Move channelAction from api/client/ -> api/urbit/
